requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
openai>=1.50.0
pypdf>=4.0.0
docx2txt>=0.8
python-dotenv>=1.0.0
streamlit>=1.38.0
